// src/configuration/cnab/interfaces/cnab-240/104/pagamento/cnab-detalhe-z-104.interface.ts
import { CnabField, CnabFieldAs } from 'src/configuration/cnab/interfaces/cnab-all/cnab-field.interface';

/**
 * PAGAMENTO DE SALÁRIOS, PAGAMENTO/CRÉDITO A FORNECEDOR E AUTOPAGAMENTO E DÉBITO AUTOMÁTICO
 *
 * @version v035 micro - 2024/09
 *
 * @extends {CnabFields}
 */
export interface CnabDetalheZ_104 {
  codigoBanco: CnabField;
  loteServico: CnabField;
  codigoRegistro: CnabFieldAs<number>;
  nsr: CnabFieldAs<number>;
  codigoSegmento: CnabFieldAs<string>;
  usoExclusivoFebraban: CnabField;
  /**
   * Z.07 - Autenticação para atender Legislação
   *
   * Número do Comprovante de Pagamento gerado após a efetivação do Débito em Conta ou Autenticação Bancária / Protocolo.
   */
  comprovanteAutenticacao: CnabField;
  /**
   * Z.08 - Ocorrências
   */
  usoExclusivoFebraban2: CnabField;
}